Move DapAbort and ProblemDetails to their own module
