+++ daphne/src/lib.rs
@@ -35,6 +35,7 @@
 //!     > requests to a collect job URI whose results have been removed.
 
 use crate::{
+    aborts::DapAbort,
     hpke::HpkeReceiverConfig,
     messages::{
         AggregationJobId, BatchId, BatchSelector, Collection, CollectionJobId,
@@ -119,270 +120,6 @@ impl From<VdafError> for DapError {
             }
         }
     }
-}
-
-/// DAP aborts.
-#[derive(Debug, thiserror::Error)]
-pub enum DapAbort {
-    /// Bad request. Sent in response to an HTTP request that couldn't be handled preoprly.
-    #[error("badRequest")]
-    BadRequest(String),
-
-    /// Invalid batch. Sent in response to a CollectReq or AggregateShareReq.
-    #[error("batchInvalid")]
-    BatchInvalid { detail: String, task_id: TaskId },
-
-    /// Batch mismatch. Sent in response to an AggregateShareReq.
-    #[error("batchMismatch")]
-    BatchMismatch { detail: String, task_id: TaskId },
-
-    /// Batch overlap. Sent in response to an CollectReq for which the Leader detects the same
-    /// Collector requesting an aggregate share which it has collected in the past.
-    #[error("batchOverlap")]
-    BatchOverlap { detail: String, task_id: TaskId },
-
-    /// Internal error.
-    #[error("internalError")]
-    Internal(#[source] Box<dyn std::error::Error + 'static + Send + Sync>),
-
-    /// Invalid batch size (either too small or too large). Sent in response to a CollectReq or
-    /// AggregateShareReq.
-    #[error("invalidBatchSize")]
-    InvalidBatchSize { detail: String, task_id: TaskId },
-
-    /// draft-wang-ppm-dap-taskprov-02: Invalid DAP task. Sent when a server opts out of a
-    /// taskprov task configuration.
-    #[error("invalidTask")]
-    InvalidTask { detail: String, task_id: TaskId },
-
-    /// Request with missing task ID.
-    #[error("missingTaskID")]
-    MissingTaskId,
-
-    /// Query mismatch. Sent in response to a CollectReq or AggregateShareReq.
-    #[error("queryMismatch")]
-    QueryMismatch { detail: String, task_id: TaskId },
-
-    /// Report rejected. Sent in response to an upload request containing a Report that the Leader
-    /// would reject during the aggregation sub-protocol.
-    #[error("reportRejected")]
-    ReportRejected { detail: String },
-
-    /// Report too late. Sent in response to an upload request for a task that is known to have
-    /// expired.
-    #[error("reportTooLate")]
-    ReportTooLate,
-
-    /// Round mismatch. The aggregators disagree on the current round of the VDAF preparation protocol.
-    /// This abort occurs during the aggregation sub-protocol.
-    #[error("roundMismatch")]
-    RoundMismatch {
-        detail: String,
-        task_id: TaskId,
-        // draft02 compatibility: The ID's definition (i.e., length in bytes) depends on which
-        // protocol is in use, hence the need for the `MetaAggregationJobId` type for representing
-        // the union of both To avoid having to propgate the lifetime parameter to `DapAbort`, we
-        // encode it right away.
-        agg_job_id_base64url: String,
-    },
-
-    /// Unauthorized HTTP request.
-    #[error("unauthorizedRequest")]
-    UnauthorizedRequest { detail: String, task_id: TaskId },
-
-    /// Unrecognized aggregation job. Sent in response to an AggregateContinueReq for which the
-    /// Helper does not recognize the indicated aggregation job.
-    #[error("unrecognizedAggregationJob")]
-    UnrecognizedAggregationJob {
-        task_id: TaskId,
-        // draft02 compatibility: The ID's definition (i.e., length in bytes) depends on which
-        // protocol is in use, hence the need for the `MetaAggregationJobId` type for representing
-        // the union of both To avoid having to propgate the lifetime parameter to `DapAbort`, we
-        // encode it right away.
-        agg_job_id_base64url: String,
-    },
-
-    /// Unrecognized message. Sent in response to a malformed or unexpected message.
-    #[error("unrecognizedMessage")]
-    UnrecognizedMessage,
-
-    /// Unrecognized DAP task. Sent in response to a request indicating an unrecognized task ID.
-    #[error("unrecognizedTask")]
-    UnrecognizedTask,
-}
-
-impl DapAbort {
-    /// Construct a problem details JSON object for this abort. `url` is the URL to which the
-    /// request was targeted and `task_id` is the associated TaskID.
-    pub fn into_problem_details(self) -> ProblemDetails {
-        let typ = format!("urn:ietf:params:ppm:dap:error:{}", self);
-        let title = self.title().to_string();
-        let (task_id, detail, agg_job_id_base64url) = match self {
-            Self::BatchInvalid { detail, task_id }
-            | Self::InvalidTask { detail, task_id }
-            | Self::BatchMismatch { detail, task_id }
-            | Self::BatchOverlap { detail, task_id }
-            | Self::InvalidBatchSize { detail, task_id }
-            | Self::QueryMismatch { detail, task_id }
-            | Self::UnauthorizedRequest { detail, task_id } => (Some(task_id), Some(detail), None),
-            Self::MissingTaskId => (
-                None,
-                Some("A task ID must be specified in the query parameter of the request.".into()),
-                None,
-            ),
-            Self::BadRequest(detail) | Self::ReportRejected { detail } => {
-                (None, Some(detail), None)
-            }
-            Self::RoundMismatch {
-                detail,
-                task_id,
-                agg_job_id_base64url,
-            } => (Some(task_id), Some(detail), Some(agg_job_id_base64url)),
-            Self::UnrecognizedAggregationJob {
-                task_id,
-                agg_job_id_base64url,
-            } => (
-                Some(task_id),
-                Some("The request indicates an aggregation job that does not exist.".into()),
-                Some(agg_job_id_base64url),
-            ),
-            Self::ReportTooLate | Self::UnrecognizedMessage | Self::UnrecognizedTask => {
-                (None, None, None)
-            }
-            Self::Internal(e) => (None, Some(e.to_string()), None),
-        };
-
-        ProblemDetails {
-            typ,
-            title,
-            task_id: task_id.map(|id| id.to_base64url()),
-            agg_job_id: agg_job_id_base64url,
-            instance: None, // TODO interop: Implement as specified.
-            detail,
-        }
-    }
-
-    /// Abort due to unexpected value for HTTP content-type header.
-    pub fn content_type<S>(req: &DapRequest<S>, expected: DapMediaType) -> Self {
-        let want_str = expected
-            .as_str_for_version(req.version)
-            .expect("could not resolve content-type for expected media type");
-
-        if let Some(got_str) = req.media_type.as_str_for_version(req.version) {
-            Self::BadRequest(format!(
-                "unexpected content-type: got {got_str}; want {want_str}"
-            ))
-        } else {
-            Self::BadRequest(format!("missing content-type: expected {want_str}"))
-        }
-    }
-
-    #[inline]
-    pub(crate) fn version_mismatch(indicated: DapVersion, expected: DapVersion) -> Self {
-        DapAbort::BadRequest(format!(
-            "DAP version of request does not match task: got {indicated:?}; want {expected:?}"
-        ))
-    }
-
-    #[inline]
-    pub(crate) fn version_unknown() -> Self {
-        DapAbort::BadRequest("DAP version of request is not recognized".into())
-    }
-
-    #[inline]
-    pub(crate) fn batch_overlap(task_id: &TaskId, batch_sel: &BatchSelector) -> Self {
-        Self::BatchOverlap {
-            detail: format!(
-                "The batch indicated by the request: {}",
-                serde_json::to_string(batch_sel).expect("failed to JSON-encode the batch selector while constructing a \"batchOverlap\" abort"),
-            ),
-            task_id: task_id.clone(),
-        }
-    }
-
-    #[inline]
-    pub(crate) fn query_mismatch(
-        task_id: &TaskId,
-        query_type_for_task: impl std::fmt::Display,
-        query_type_for_request: impl std::fmt::Display,
-    ) -> Self {
-        Self::QueryMismatch {
-            detail: format!("The task's query type is \"{query_type_for_task}\", but the request indicates \"{query_type_for_request}\"."),
-            task_id: task_id.clone(),
-        }
-    }
-
-    #[inline]
-    pub fn report_rejected(failure_reason: TransitionFailure) -> Self {
-        let detail = match failure_reason {
-            TransitionFailure::BatchCollected => {
-                "The report pertains to a batch that has already been collected."
-            }
-            TransitionFailure::ReportReplayed => {
-                "A report with the same ID was uploaded previously."
-            }
-            _ => return DapError::Fatal(format!("Attempted to construct a \"reportRejected\" abort with unexpected transition failure: {failure_reason:?}")).into(),
-        };
-
-        Self::ReportRejected {
-            detail: detail.into(),
-        }
-    }
-
-    fn title(&self) -> &'static str {
-        match self {
-            Self::BatchInvalid { .. } => "Batch boundary check failed",
-            Self::BatchMismatch { .. } => "Aggregators disagree on the set of reports in the batch",
-            Self::BatchOverlap { .. } => "The selected batch overlaps with a previous batch",
-            Self::InvalidBatchSize { .. } => "Batch size is invalid",
-            Self::InvalidTask { .. } => "Opted out of Taskprov task",
-            Self::QueryMismatch { .. } => "Query type does not match the task",
-            Self::RoundMismatch { .. } => "Aggregation round indicated by peer does not match host",
-            Self::MissingTaskId => "Request for HPKE configuration with unspecified task",
-            Self::ReportRejected { .. } => "Report rejected",
-            Self::ReportTooLate => "The requested task expires after report timestamp",
-            Self::UnauthorizedRequest { .. } => "Request authorization failed",
-            Self::UnrecognizedAggregationJob { .. } => "Unrecognized aggregation job",
-            Self::UnrecognizedMessage => "Failed to parse the request body",
-            Self::UnrecognizedTask => "Task indicated by request is not recognized",
-            Self::BadRequest(..) => "Invalid request",
-            Self::Internal(..) => "Internal server error",
-        }
-    }
-}
-
-impl From<DapError> for DapAbort {
-    fn from(e: DapError) -> Self {
-        match e {
-            e @ DapError::Fatal(..) => Self::Internal(Box::new(e)),
-            DapError::Abort(abort) => abort,
-            DapError::Transition(failure_reason) => Self::report_rejected(failure_reason),
-        }
-    }
-}
-
-impl From<CodecError> for DapAbort {
-    fn from(_e: CodecError) -> Self {
-        Self::UnrecognizedMessage
-    }
-}
-
-/// A problem details document compatible with RFC 7807.
-#[derive(Debug, Deserialize, Serialize)]
-pub struct ProblemDetails {
-    #[serde(rename = "type")]
-    pub typ: String,
-    pub title: String,
-    #[serde(rename = "taskid")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) task_id: Option<String>,
-    #[serde(rename = "aggregationjobid")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) agg_job_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) instance: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub detail: Option<String>,
 }
 
 /// DAP version used for a task.
@@ -1173,10 +910,11 @@ impl MetaAggregationJobId<'_> {
     }
 }
 
+pub mod aborts;
 pub mod auth;
 pub mod constants;
 #[cfg(test)]
-pub mod constants_test;
+mod constants_test;
 pub mod hpke;
 #[cfg(test)]
 mod hpke_test;
